app/requirements.txt
alembic==1.5.7
amqp==5.0.5
astroid==2.4.2
atomicwrites==1.4.0
attrs==20.2.0
Babel==2.9.0
billiard==3.6.3.0
blinker==1.4
celery==5.0.5
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
cloudinary==1.23.0
colorama==0.4.4
Flask==1.1.2
Flask-Admin==1.5.7
Flask-BabelEx==0.9.4
Flask-Cors==3.0.9
Flask-JWT-Extended==3.24.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==2.7.0
flask-mongoengine==0.9.5
Flask-Principal==0.4.0
Flask-Script==2.0.6
Flask-Security==3.0.0
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
gunicorn==20.0.4
idna==2.10
importlib-metadata==2.0.0
iniconfig==1.1.1
isort==5.6.4
itsdangerous==1.1.0
Jinja2==2.11.2
kombu==5.0.2
lazy-object-proxy==1.4.3
Mako==1.1.4
MarkupSafe==1.1.1
mccabe==0.6.1
mongoengine==0.20.0
mysqlclient==2.0.1
numpy==1.19.4
packaging==20.4
passlib==1.7.4
pika==1.2.0
pluggy==0.13.1
prompt-toolkit==3.0.17
psycopg2==2.8.6
py==1.9.0
PyJWT==1.7.1
pylint==2.6.0
pymongo==3.11.0
pyparsing==2.4.7
pytest==6.1.1
python-dateutil==2.8.1
python-dotenv==0.14.0
python-editor==1.0.4
pytz==2020.4
redis==3.5.3
requests==2.24.0
six==1.15.0
speaklater==1.3
SQLAlchemy==1.3.20
toml==0.10.1
urllib3==1.25.11
vine==5.0.0
wcwidth==0.2.5
Werkzeug==1.0.1
wrapt==1.12.1
WTForms==2.3.3
zipp==3.3.1
